+++ bin/generateAppConfig.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+
+require('../scripts/generate-app-config/generateAppConfigCli');