+++ dev-site-config/navigation.config.js
@@ -1,0 +1,16 @@
+const navConfig = {
+  navigation: {
+    index: '/home',
+    links: [{
+      path: '/home',
+      text: 'Home',
+      pageTypes: ['home'],
+    }, {
+      path: '/getting-started',
+      text: 'Getting Started',
+      pageTypes: ['doc'],
+    }],
+  },
+};
+
+module.exports = navConfig;