+++ dev-site-config/site.config.js
@@ -1,0 +1,12 @@
+const navConfig = require('./navigation.config');
+
+const siteConfig = {
+  navConfig,
+
+  appConfig: {
+    locales: ['en'],
+    bidirectional: false,
+  },
+};
+
+module.exports = siteConfig;