+++ src/app/ConfigureUtilities.jsx
@@ -1,24 +1,3 @@
-import React from 'react';
-import IconSettings from 'terra-icon/lib/icon/IconSettings';
-
-const generateItemConfig = (defaultItem, items, key) => {
-  const childKeys = {};
-  items.forEach((item) => {
-    childKeys[item] = {
-      key: item,
-      title: item,
-      isSelected: defaultItem === item,
-      isSelectable: true,
-    };
-  });
-
-  return {
-    key,
-    title: `${key}: ${defaultItem}`,
-    childKeys,
-  };
-};
-
 const convertMenuItemToArray = (menuItem) => {
   const updatedItem = Object.assign({}, menuItem);
   let menuItems = [];
@@ -51,48 +30,19 @@ const updateSelectedItem = (parentMenuItem, selectedItem) => {
 };
 
 class ConfigureUtilties {
-
-  static generateInitialUtiltiesConfig(appConfig) {
-    const hasThemes = appConfig.themes && Object.keys(appConfig.themes).length > 0;
-    const hasLocals = appConfig.locales && appConfig.locales.length > 0;
-
-    if (!(hasThemes || hasLocals || appConfig.bidirectional)) {
+  static convertChildkeysToArray(config) {
+    if (config === undefined) {
       return undefined;
     }
-
-    const rootMenuChildKeys = {};
-
-    if (hasThemes) {
-      rootMenuChildKeys.Theme = generateItemConfig(appConfig.defaultTheme, Object.keys(appConfig.themes), 'Theme');
-    }
-
-    if (hasLocals) {
-      rootMenuChildKeys.Locale = generateItemConfig(appConfig.defaultLocale, appConfig.locales, 'Locale');
-    }
-
-    if (appConfig.bidirectional) {
-      rootMenuChildKeys.Bidi = generateItemConfig(appConfig.defaultDir, ['ltr', 'rtl'], 'Bidi');
-    }
-
-    return {
-      accessory: <IconSettings />,
-      menuItems: {
-        key: 'menu',
-        title: 'Config',
-        childKeys: rootMenuChildKeys,
-      },
-      initialSelectedKey: 'menu',
-      onChange: (event, { key, metaData }) => { metaData.onChange(key); },
-    };
-  }
-
-  static convertChildkeysToArray(config) {
     const updatedConfig = Object.assign({}, config);
     updatedConfig.menuItems = convertMenuItemToArray(updatedConfig.menuItems);
     return updatedConfig;
   }
 
   static addCallbackFunctions(config, metaData = {}) {
+    if (config === undefined) {
+      return undefined;
+    }
     const updatedConfig = Object.assign({}, config);
     updateMetaData(updatedConfig.menuItems.childKeys.Theme, metaData.Theme);
     updateMetaData(updatedConfig.menuItems.childKeys.Locale, metaData.Locale);
@@ -101,6 +51,9 @@ class ConfigureUtilties {
   }
 
   static updateSelectedItems(config, theme, locale, dir) {
+    if (config === undefined) {
+      return undefined;
+    }
     const updatedConfig = Object.assign({}, config);
     updateSelectedItem(updatedConfig.menuItems.childKeys.Theme, theme);
     updateSelectedItem(updatedConfig.menuItems.childKeys.Locale, locale);
@@ -109,4 +62,5 @@ class ConfigureUtilties {
     return updatedConfig;
   }
 }
+
 export default ConfigureUtilties;