+++ tests/app/MockComponent.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const MockComponent = () => <div>Mock Component</div>;
-
-export default MockComponent;